// restore/lib/types.ts
export interface CompositionItem {
  productId: string;
  quantity: number;
  name?: string;
  price?: number;
}

export interface Departamento {
  id: string;
  codigo?: string;
  nome: string;
  ativo: boolean;
}

export interface ExpenseCategory {
  id: string;
  nome: string;
}

export interface Categoria {
  id: string;
  codigo?: string;
  nome: string;
  departamento_id: string;
}

export interface Subcategoria {
  id: string;
  codigo?: string;
  nome: string;
  categoria_id: string;
}

export interface Product {
  id: string;
  companyId: string;
  name: string;
  sku: string;
  costPrice: number;
  salePrice: number;
  stock: number;
  minStock: number;
  image: string;
  brand?: string;
  unit?: string;
  supplier?: string;
  profit?: number;
  profitPercentage?: number;
  composition?: CompositionItem[];
  status?: 'Ativo' | 'Inativo';
  subcategoria_id?: string;
  codigo_mercadologico?: string;
  category?: string;
  subgroup?: string;
  validade?: string;
  has_had_stock?: boolean;
}